int -> long for setTelephoneNumber

+++ src/Q6/Person.java
@@ -33,7 +33,7 @@ public class Person {
         return telephoneNumber;
     }
 
-    public void setTelephoneNumber(int telephoneNumber) {
+    public void setTelephoneNumber(long telephoneNumber) {
         this.telephoneNumber = telephoneNumber;
     }
 }
